requirements.txt
Flask==2.2.5
requests==2.31.0
python-dotenv==1.0.0
# openai==0.28.0  # optional, uncomment if you enable real API calls
